+++ src/main/scala/com/nec/ve/VeRDD.scala
@@ -1,0 +1,25 @@
+package com.nec.ve
+
+import com.nec.ve.VeColBatch.VeColVector
+import org.apache.spark.rdd.RDD
+
+object VeRDD {
+  def exchange(rdd: RDD[(Int, VeColVector)])(implicit veProcess: VeProcess): RDD[VeColVector] =
+    rdd
+      .mapPartitions(
+        f = iter =>
+          iter.map { case (p, v) =>
+            (p, (v, v.serialize()))
+          },
+        preservesPartitioning = true
+      )
+      .sortByKey()
+      .mapPartitions(
+        f = iter => iter.map { case (_, (v, ba)) => v.deserialize(ba) },
+        preservesPartitioning = true
+      )
+
+  implicit class RichKeyedRDD(rdd: RDD[(Int, VeColVector)]) {
+    def exchangeBetweenVEs()(implicit veProcess: VeProcess): RDD[VeColVector] = exchange(rdd)
+  }
+}
